fix: fix unlock oma conflict (#606)
`ask_continue_no_use_battery` input ctrl-c will use `unlock_oma`,
and ctrlc `signal_handler` again to use `unlock_oma`, this can cause
deletion conflicts

+++ src/main.rs
@@ -655,7 +655,9 @@ pub fn lock_oma_inner(sysroot: &Path) -> io::Result<()> {
 
 /// Unlock oma
 pub fn unlock_oma() -> io::Result<()> {
-    if let Some(lock) = LOCK.get() {
+    if let Some(lock) = LOCK.get()
+        && lock.exists()
+    {
         std::fs::remove_file(lock)?;
     }
